Increased Graphics Cap!
- This commit increases the overworld graphics cap from 256 to 512.
- **YOU WILL NEED TO REBUILD TOOLS AND CLEAN BUILD THE ROM!**
- Lots of changing u8s to u16s.
- The main change is in global.fieldmap.h, where the EventObjectTemplate 
was changed to handle a u16, and the EventObject was change to handle a 
u9 for the graphicsId. (The compiler complained about making spriteId a 
u6, because it can't hold the value 64).
- All maps are affected. Mapjson will need to be recompiled, and all 
maps will need to be recompiled, or you will see Brendan everywhere.
- Old save files will load events incorrectly, but all should be fine 
with a map change.

By https://github.com/tustin2121

+++ src/field_specials.c
@@ -98,7 +98,7 @@ void SetPlayerGotFirstFans(void);
 u16 GetNumFansOfPlayerInTrainerFanClub(void);
 
 static void RecordCyclingRoadResults(u32, u8);
-static void LoadLinkPartnerEventObjectSpritePalette(u8 graphicsId, u8 localEventId, u8 paletteNum);
+static void LoadLinkPartnerEventObjectSpritePalette(u16 graphicsId, u8 localEventId, u8 paletteNum);
 static void Task_PetalburgGymSlideOpenRoomDoors(u8 taskId);
 static void PetalburgGymSetDoorMetatiles(u8 roomNumber, u16 metatileId);
 static void Task_PCTurnOnEffect(u8);
@@ -545,7 +545,7 @@ void SpawnLinkPartnerEventObject(void)
     };
     u8 myLinkPlayerNumber;
     u8 playerFacingDirection;
-    u8 linkSpriteId;
+    u16 linkSpriteId;
     u8 i;
 
     myLinkPlayerNumber = GetMultiplayerId();
@@ -608,7 +608,7 @@ void SpawnLinkPartnerEventObject(void)
     }
 }
 
-static void LoadLinkPartnerEventObjectSpritePalette(u8 graphicsId, u8 localEventId, u8 paletteNum)
+static void LoadLinkPartnerEventObjectSpritePalette(u16 graphicsId, u8 localEventId, u8 paletteNum)
 {
     u8 adjustedPaletteNum;
     // Note: This temp var is necessary; paletteNum += 6 doesn't match.
